requirements.txt
langchain
langchain-community
langchain-core
langchain-text-splitters

chromadb
sentence-transformers

python-dotenv
